Add lazy-seq regression unit test
We need to make sure that we do not pass a lazy-seq in to sdo. This will
cause some rules to not be loaded if a large enough number of rules are
specified to load.

This bug has occurred twice. Add a unit test to ensure it doesn't happen
a third.