Create custom lein task for coveralls.io
The coveralls.io support in the cloverage plugin is rather weak. It
doesn't extract git information, PR information, or the repo token. It
also uses the wrong key to identify circle-ci build numbers (that one
isn't cloverage's fault -- that's coverall's fault for having an
inconsistent API).

This task adds the missing fields to the json and then submits it to
coveralls.io.